Auto-update the referenced Glean version when GeckoView updates
A-C still references the exact Glean version.
Glean is now shipped as part of the geckoview-omni package.
We can auto-update that version as part of the automated GeckoView
update.

This also changes the GeckoView version checks from the `geckoview`
package to the `geckoview-omni` package.
A-C switched to the omni package in September 2021[1]

[1]: https://github.com/mozilla-mobile/android-components/commit/0b349f48c91a50bb7b4ffbf40c6c122ed18142d3